backend/requirements.txt
fastapi==0.112.0
uvicorn==0.30.6
groq==0.9.0
langchain==0.2.16
langchain-groq==0.1.9
langchain-community==0.2.16
supabase==2.7.4
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
